pyproject.toml
[project]
name = "faq-chatbot"
version = "0.1.0"
description = "FAQ Chatbot with Clean Architecture using Streamlit and Milvus"
authors = [
    {name = "Your Name",email = "your.email@example.com"}
]
readme = "README.md"
package-mode = false


[tool.poetry.dependencies]
python = ">=3.9,<3.9.7 || >3.9.7,<4.0"
streamlit = "^1.47.0"
fastapi = "^0.104.1"
uvicorn = "^0.24.0"
pymilvus = "^2.3.4"
sentence-transformers = "^2.2.2"
pydantic = "^2.5.0"
python-multipart = "^0.0.6"
openai = "^1.3.0"
python-dotenv = "^1.0.0"
PyPDF2 = "^3.0.1"
langchain = "^0.3.26"
langchain-openai = "^0.2.0"
langgraph = "^0.5.3"
langsmith = "^0.3.45"


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
